Reduce amount of seeks and reads on pagemap data
Instead of seeking and reading single pagemap details for each page
we slurp up as much data as we can in one read and only when we
hit a memory map boundary do we need to seek and read the next
chunk of pagepage data.  This reduced a typical load from 35% down
to 8% on my laptop.

Signed-off-by: Colin Ian King <colin.king@canonical.com>