Fix potential integer overflow on 32 bit multiplications
Static analysis is warning that the 32 bit multiplications when
calculating a 64 bit data and page index can potentially overflow.
While this is unlikely that anyone will have a x,y cursor large
enough (e.g. x * y > 32 bits) to cause overflow it's sensible to
cast these to the same type as the 64 bit indexes to avoid any
overflow warnings if the tool chain gets smarter in the future.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>